bin: add alias to all greybus modules
Add links to a new script that can be used to load the manifest of any
of the greybus protocols. So, after login if you run sdio will do all
the necessary work to start gbsim and load the sdio manifest.

At any time you can run another greybus protocol name and will replace
the previous running without the need to restart.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>