package: add gbutils package
To avoid the need to always sync lsgb every time there is
change, add a new package that will clone the github projectara
gb-utils repository and will install lsgb to the target
rootfs.

Remove also the previous version of lsgb in the overlay section.

By default it will clone the github repo, however the makefile
include some commented lines that will allow to use a local
clone instead for development.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>