build: adapt to buildroot 2016.11
Current head of buildroot, version 2016.11, made a change that needs some
adaptation to external repositories. This change is to allow to have
more than one external at the time, but for that to happen each external
needs to add a descriptor (external.desc) with a unique name and adjust
the all paths references to BR_EXTERNAL to BR_EXTERNAL_ARA_PATH.

This is what is done in this patch.

Signed-off-by: Rui Miguel Silva <rmfrfs@gmail.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>