kernel: use staging next as kernel version
Now that the greybus code is in the staging kernel driver area, we move
and start using that tree. This allow us to remove reference to external
greybus dir (GBDIR) and make adjustments to scripts, add configs files
and update README documentation (please read it again for the changes in
the rebuild process).

Signed-off-by: Rui Miguel Silva <rmfrfs@gmail.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>