package: add greybus_loopback_test package
The loopback_test utility allows various different types of bandwidth and
latency tests to be performed against a greybus loopback Interface. This
utility is available in the kernel tree as loopback_test, this package is
called greybus_loopback_test here to fit in with the other package names in
the greybus overlay.

Once built this package will build the utility in
drivers/staging/greybus/tools and install it to /usr/sbin/loopback_test in
the target buildroot filesystem.

Signed-off-by: Bryan O'Donoghue <pure.logic@nexus-software.ie>
Signed-off-by: Rui Miguel Silva <rmfrfs@gmail.com>